# feat(sha2): implement intrinsic based version for sha256