--- a/sha2/Cargo.toml
+++ b/sha2/Cargo.toml
@@ -21,6 +21,13 @@ opaque-debug = "0.2"
 sha2-asm = { version = "0.5", optional = true }
 libc = { version = "0.2.68", optional = true }
 
+[dependencies.lazy_static]
+version = "1.4.0"
+default-features = false
+# no_std feature is an anti-pattern. Why, lazy_static, why?
+# See https://github.com/rust-lang-nursery/lazy-static.rs/issues/150
+features = ["spin_no_std"]
+
 [dev-dependencies]
 digest = { version = "0.9", features = ["dev"] }
 hex-literal = "0.2"