--- a/sha2/Cargo.toml
+++ b/sha2/Cargo.toml
@@ -21,6 +21,9 @@ opaque-debug = "0.2"
 sha2-asm = { version = "0.5", optional = true }
 libc = { version = "0.2.68", optional = true }
 
+[target.'cfg(any(target_arch = "x86", target_arch = "x86_64"))'.dependencies]
+raw-cpuid = "7.0.3"
+
 [dependencies.lazy_static]
 version = "1.4.0"
 default-features = false