--- a/sha2/src/lib.rs
+++ b/sha2/src/lib.rs
@@ -93,7 +93,6 @@ mod platform;
 mod sha256;
 #[cfg(any(target_arch = "x86", target_arch = "x86_64"))]
 mod sha256_intrinsics;
-#[cfg(any(not(feature = "asm"), feature = "asm-aarch64"))]
 mod sha256_utils;
 mod sha512;
 #[cfg(any(not(feature = "asm"), target_arch = "aarch64", feature = "compress"))]