@dignifiedquire 2019-11-05
I have also added feature detection at runtime, based on how blake2_simd implements it (https://github.com/oconnor663/blake2_simd/blob/master/blake2s/src/guts.rs) 

As otherwise it gets tricky to deploy this without creating many different builds