@newpavlov 2019-11-29
This is why I wanted to add those intrinsics only in v0.9. For now I think we can use a disabled by default feature or wait until I will get time to work on RustCrypto/traits#43 (probably only in January).