@tarcieri 2020-06-10
@dignifiedquire can you rebase? we've bumped MSRV to 1.41+