@Fleshgrinder 2020-06-10
@dignifiedquire are you planning to implement the same for the other SHA algos?