@newpavlov 2020-06-11
@Fleshgrinder 
See #166. x86 SHA extension only covers SHA-1 and SHA-256.