@dignifiedquire 2020-06-11
I have rebased the code and pulled in the changes I had in https://github.com/filecoin-project/rust-sha2ni 

Fyi rust-sha2ni just got a security audit through NCC (the sha2-256 part), will publishe the report soon.