@newpavlov 2020-06-11
@dignifiedquire 
I've drafted the alternative PR partially using yours as a reference. I think it has turned out quite nice and compact. Also it should have the same performance as code in this PR. Can you please check it out?