@newpavlov 2020-06-15
#167 is merged, so I will close this PR.